Coucou

+++ Sources/Memory/ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: ahamouda <ahamouda@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/07/05 04:43:00 by ahamouda          #+#    #+#             */
-/*   Updated: 2016/09/09 15:22:08 by ahamouda         ###   ########.fr       */
+/*   Updated: 2016/09/16 11:54:53 by ahamouda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	*ft_memchr(const void *s, int c, size_t n)
 	while (p < n)
 	{
 		if (x[p] == (const unsigned char)c)
-			return ((void*)((unsigned long)x + (unsigned long)p));
+			return ((void*)((uintptr_t)x + (uintptr_t)p));
 		p++;
 	}
 	return (0);
